// src/shared/enums/error-codes.enum.ts
export enum ErrorCodesEnum {
    AMOUNT_IS_EMPTY = 'amount_is_empty',
    AMOUNT_IS_INVALID = 'amount_is_invalid',
    BALANCE_FORMAT_IS_INVALID = 'balance_format_is_invalid',
    BIRTH_DATE_IS_EMPTY = 'birth_date_is_empty',
    BIRTH_DATE_IS_INVALID = 'birth_date_is_invalid',
    EMAIL_IS_EMPTY = 'email_is_empty',
    EMAIL_IS_INVALID = 'email_is_invalid',
    EMAIL_IS_UNAVAILABLE = 'email_is_unavailable',
    EMAIL_OR_PASSWORD_IS_INVALID = 'email_or_password_is_invalid',
    ID_HAS_INVALID_FORMAT = 'id_has_invalid_format',
    ID_IS_EMPTY = 'id_is_empty',
    INSUFFICIENT_BALANCE = 'insufficient_balance',
    INTERNAL_ERROR = 'internal_error',
    NAME_IS_EMPTY = 'name_is_empty',
    UNKNOW_ERROR = 'unknow_error',
    USER_NOT_FOUND = 'user_not_found',
    UNAUTHORIZED = 'unauthorized',
    WALLET_NOT_FOUND = 'wallet_not_found',
    PASSWORD_IS_EMPTY = 'password_is_empty',
    USER_ALREADY_HAS_CREDENTIALS = 'user_already_has_credentials',
}