// src/shared/enums/error-messages.enum.ts
export enum ErrorMessagesEnum {
    AMOUNT_IS_EMPTY = 'O valor não pode ser vazio.',
    AMOUNT_IS_INVALID = 'O valor deve ser maior que zero.',
    BALANCE_FORMAT_IS_INVALID = 'O formato do saldo é inválido',
    BIRTH_DATE_IS_EMPTY = 'A data de nascimento não pode ser vazio.',
    BIRTH_DATE_IS_INVALID = 'A data de nascimento deve ser uma data válida.',
    EMAIL_IS_EMPTY = 'O email não pode ser vazio.',
    EMAIL_IS_INVALID = 'O email deve ser um email válido.',
    EMAIL_IS_UNAVAILABLE = 'Este email está indisponível.',
    ID_HAS_INVALID_FORMAT = 'O formato do id é inválido.',
    ID_IS_EMPTY = 'O id não pode ser vazio.',
    INSUFFICIENT_BALANCE = 'Saldo insuficiente para realizar transferência.',
    INTERNAL_ERROR = 'Erro interno no servidor.',
    NAME_IS_EMPTY = 'O nome não pode ser vazio.',
    UNKNOW_ERROR = 'Erro desconhecido.',
    USER_NOT_FOUND = 'Usuário não encontrado.',
    EMAIL_OR_PASSWORD_IS_INVALID = 'Email ou senha está inválido.',
    UNAUTHORIZED = 'Unauthorized.',
    WALLET_NOT_FOUND = 'Carteira não encontrada.',
    PASSWORD_IS_EMPTY = 'A senha não pode ser vazia.',
    USER_ALREADY_HAS_CREDENTIALS = 'Usuário já possui credenciais cadastradas.',
}